refactor(security): extract suspicious-origin detection into internal/security/netflag
The brief hypothesised that security_credentials* and security_flagging* were one
secret-scanning cluster. They are not. There is not a single call edge between
them, they emit different types (SecurityFinding vs capture.SecurityFlag), and
they are documented under different features. flagging is a self-contained leaf
whose only entry point is analyzeNetworkSecurity, consumed by one file
(security_network_findings.go); credentials is bound to *SecurityScanner methods
and shares redactSecret with the PII check. So flagging extracts cleanly and
credentials cannot — see the PR description for the full evidence.

analyzeNetworkSecurity becomes netflag.Analyze; every detector stays unexported.

Also drops the //go:build integration tag from the detector tests. Its stated
reason — "NetworkWaterfallEntry, NewCapture, SecurityFlag aren't available" — is
stale; those types are exported from internal/capture and the tests compile and
pass in a normal run. Twenty-one tests that were being skipped now run by
default, taking netflag to 95.6% statement coverage.
