+++ internal/security/netflag/detectors.go
@@ -1,6 +1,7 @@
+// detectors.go — Individual TLD, port, IP-origin, typosquat and mixed-content detectors.
 // Purpose: Detects suspicious TLDs, typosquatting, and anomalous network patterns.
 // Why: Separates detection heuristics from static data and distance calculations.
-package security
+package netflag
 
 import (
 	"net"
