+++ internal/security/netflag/netflag.go
@@ -1,16 +1,23 @@
+// netflag.go — Runs every origin/resource detector for one captured network entry.
 // Purpose: Flags suspicious origins, ports, and supply-chain indicators from captured network activity.
 // Why: Surfaces high-signal threat indicators that are otherwise easy to miss in raw telemetry.
 // Docs: docs/features/feature/security-hardening/index.md
 
-package security
+// Package netflag detects suspicious network origins — abusive TLDs, non-standard
+// ports, raw IP origins, typosquatted CDN domains and mixed content — and reports
+// them as capture.SecurityFlag values.
+//
+// It is a leaf package: it depends only on capture/util and never on a sibling
+// security package. Callers translate its flags into their own finding shapes.
+package netflag
 
 import (
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/capture"
 	"github.com/brennhill/Kaboom-Browser-AI-Devtools-MCP/internal/util"
 )
 
-// analyzeNetworkSecurity runs the full set of origin/resource checks for one network entry.
-func analyzeNetworkSecurity(entry capture.NetworkWaterfallEntry, pageURL string) []capture.SecurityFlag {
+// Analyze runs the full set of origin/resource checks for one network entry.
+func Analyze(entry capture.NetworkWaterfallEntry, pageURL string) []capture.SecurityFlag {
 	origin := util.ExtractOrigin(entry.URL)
 	if origin == "" {
 		return nil
